Move promise module
